requirements.txt
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.0
pandas==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.1
gunicorn==21.2.0
jinja2
fastapi
python-multipart
uvicorn 